keycloak-config/requirements.txt
python-keycloak==0.22.0
